FixDatabase And Price Types

+++ Models/CarPart.cs
@@ -10,8 +10,7 @@ namespace coursa4.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public int QuantityInStock { get; set; }
-
     }
 }
